% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-charnum-cols.R
\name{append_charnum_cols}
\alias{append_charnum_cols}
\title{Character and numeric columns}
\usage{
append_charnum_cols(x, cols, old = "_c", new = "_n", beside = TRUE)
}
\arguments{
\item{x}{A data.frame}

\item{cols}{A vector of column names to be transformed from character to
numeric.}

\item{old, new}{A suffix added to the old/new columns; NULL will not change
name; these cannot be the same}

\item{beside}{Logical, if TRUE will add new columns beside old; otherwise
will move to the end.}
}
\description{
Append data.frame with character and numeric columns
}
