% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dirty}
\alias{dirty}
\title{Example dirty data}
\format{
A data frame with x rows and y variables
\describe{
\item{V1}{variable 1, description}
\item{V2}{variable 2, description}
\item{V3}{variable 3, description}
\item{V4}{variable 4, description}
\item{V5}{variable 5, description}
}
}
\usage{
dirty
}
\description{
A test data set
}
\keyword{datasets}
