% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{earliest_date}
\alias{earliest_date}
\title{Inexact dates}
\usage{
earliest_date(year, month = NULL, day = NULL)
}
\arguments{
\item{year}{Year}

\item{month}{Month}
}
\description{
Dates sometimes suck
}
\examples{
earliest_date(2019, 1, 0)
}
