% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-factor-seq.R
\name{expand_factor_seq}
\alias{expand_factor_seq}
\title{Expand ordered factor sequence}
\usage{
expand_factor_seq(x, min_level = min(x), by = 1L)
}
\arguments{
\item{x}{A vector of ordered factors}

\item{min_level}{Character.  The minimal for the sequence}

\item{by}{Integer. Increment of the sequence; 1L or greater.}
}
\description{
Outputs a vector of
}
\examples{
x <- factor(letters[c(5:15, 20)], levels = letters, ordered = TRUE)
expand_factor_seq(x)
expand_factor_seq(x, "j", by = 3L)
}
