% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-na.R
\name{keep_na}
\alias{keep_na}
\alias{keep_nin}
\title{Filter for missing data}
\usage{
keep_na(df, cols = NULL)

keep_nin(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{Columns to include.  If NULL, all columns will be examined.}
}
\description{
Filter rows for missing data
}
