% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{na_data}
\alias{na_data}
\title{Example data with NAs}
\format{
A data frame with x rows and y variables
\describe{
\item{logical}{logical values}
\item{character}{character values}
\item{integer}{integer values}
\item{double}{double values}
\item{date}{date values}
\item{infinity}{infinity values}
}
}
\usage{
na_data
}
\description{
A test data set
}
\keyword{datasets}
