% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-data-frames.R
\name{qc_data_frame}
\alias{qc_data_frame}
\title{"QC" data frames}
\usage{
qc_data_frame(target, reference, index, keep_all = FALSE, ...)
}
\arguments{
\item{target}{Targer (test) table}

\item{reference}{Reference table}

\item{index}{The name or the index}

\item{keep_all}{Logical, passed to `reindex()` (see details)}

\item{...}{additional arguments passed to [stringdist::stringdist]}
}
\value{
A data.frame with the index values and differences in target and comparisons.
}
\description{
Compare values of two data.frames base on an index
}
\details{
All values from the target and reference will be returned as characters.
`keep_all` is passed to the `keep_all` argument in `reindex()`.
If this is `TRUE`,
}
