% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-data-frames.R
\name{qc_data_frames}
\alias{qc_data_frames}
\title{"QC" data frames}
\usage{
qc_data_frames(reference, target, index, string_distances = FALSE, ...)
}
\arguments{
\item{reference, target}{tables to join}

\item{index}{The name or the index}

\item{string_distances}{Logical.  If true, distances will be calculated.}

\item{...}{additional arguments passed to [stringdist::stringdist]}
}
\value{
A data.frame with the index values and differences in target and comparisons.
}
\description{
Compare values of two data.frames base on an index
}
\details{
All values from the target and reference will be returned as characters.
}
