% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindex.R
\name{reindex}
\alias{reindex}
\title{Re-index}
\usage{
reindex(df, index, new_index)
}
\arguments{
\item{df}{A data.frame}

\item{index}{The name of the column to reindex by}

\item{new_index}{the new order of the index}
}
\description{
Re-index a column based on a new index
}
\examples{
iris1 <- head(iris, 5)
iris1$index <- 1:5
iris2 <- iris1[c(2, 3, 5, 4, 1), ]
reindex(iris1, "index", iris2$index)
}
