% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-names.R
\name{remove_names}
\alias{remove_names}
\title{Remove names from an object}
\usage{
remove_names(object)
}
\arguments{
\item{object}{An object}
}
\description{
Removes the `names`, `dimnames` and/or `rownames` from an object.
Documentation for [base::unname()], this will not throw an error when used with a dataframe.
For `data.frames`, the `rownames()`, default to the row numbers.
}
\examples{
df <- head(iris)
rownames(df) <- letters[1:6]
unname(df)               ## removes column names
\dontrun{
unname(df, force = TRUE) ## doesn't remove rownames; throws an error
}
remove_names(df)         ## removes column and row names
}
