% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{split_date}
\alias{split_date}
\alias{parse_date}
\title{Parses dates}
\usage{
split_date(
  x,
  year = "year",
  month = "month",
  day = "day",
  dates_to_row = FALSE
)

parse_date(x, cols, year = "year", month = "month", day = "day", sep = "_")
}
\arguments{
\item{x}{A vector or data.frame}

\item{year}{Name for year column or column suffix}

\item{month}{Name for month column or column suffix}

\item{day}{Name for day column or column suffix}

\item{dates_to_row}{Logical, if `TRUE`, adds dates to the row nmaes}

\item{cols}{A character vector of the columns to parse into dates}

\item{sep}{String to use to separate new columns}
}
\description{
Separates dates from a vector or a data.frame
}
\examples{
x <- c("2010-01-12", "2020-09-30", "1999-12-31")
split_date(x)

xx <- data.frame(
  x1 = 1:3,
  x2 = runif(3),
  date1 = as.Date(c("1950-10-05", "2020-04-29", "1992-12-17")),
  x3 = letters[1:3],
  date2 = as.Date(c("2010-01-12", "2020-09-30", "1999-12-31")))
parse_date(xx, c("date1", "date2"))
}
