% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data_target}
\alias{test_data_target}
\title{Test data set for QC (target)}
\format{
A data frame with x rows and y variables
\describe{
\item{index}{index values}
\item{character}{character values}
\item{factor}{factor values}
\item{ordered}{ordered values}
\item{integer}{integer values}
\item{double}{double values}
}
}
\usage{
test_data_target
}
\description{
A test data set for QC - target set
}
\keyword{datasets}
