% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which-na.R
\name{which_na}
\alias{which_na}
\alias{which_na_inf}
\alias{which_min_na}
\title{Which NA functions}
\usage{
which_na(x)

which_na_inf(x)

which_min_na(x)

which_min_na(x)
}
\arguments{
\item{x}{a vector of values}
}
\description{
Which wrappers for finding NA values
}
