% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-na.R
\name{get_na}
\alias{get_na}
\alias{get_na_inf}
\title{Get NA values}
\usage{
get_na(.data, vars, ...)

get_na_inf(.data, ...)
}
\arguments{
\item{.data}{A data.frame.}

\item{vars}{Column names of the data set to check on}

\item{...}{Additional arguments passed to methods}
}
\description{
Return NA values from a data set.
}
