% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{keep_na}
\alias{keep_na}
\alias{keep_nin}
\title{Filter for missing data}
\usage{
keep_na(df, ...)

keep_nin(df, ...)
}
\arguments{
\item{df}{A data.frame.}

\item{...}{Columns to include.  If not included, all columns filtered on.}
}
\description{
Filter rows for missing data
}
\details{
`...` supports tidy select
}
\examples{
\dontrun{
load(dirtyr)
keep_na(dirty, V8)
keep_nin(dirty, V4)
}
}
