% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-na.R
\name{nas}
\alias{nas}
\alias{is_na_inf}
\alias{count_na}
\alias{count_na_vars}
\title{Additional na functions}
\usage{
is_na_inf(x)

count_na(x)

count_na_vars(df)
}
\arguments{
\item{x}{a vector of values}
}
\description{
Additional functions to sort for NA values
}
