% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{possible_date}
\alias{possible_date}
\alias{earliest_date}
\alias{latest_date}
\alias{unknown_date}
\title{Inexact dates}
\usage{
earliest_date(year, month = NULL, day = NULL)

latest_date(year, month = NULL, day = NULL)

unknown_date(x, format = "ymd", possible = c("earliest", "latest"))
}
\arguments{
\item{year}{Year}

\item{month}{Month}

\item{day}{Day}

\item{x}{a date "string" (see details)}

\item{format}{A date "format" (see details)}

\item{possible}{Whether to look at earliest or latest}
}
\description{
Incomplete dates cannot be directly coerced into date values in R.
These functions help
}
\details{
`unknown_date` is a vectorized function which implements the non-vectorized
  `earliest_date` and `latest_date`.
If a split vector of 2 is passed to `unknown_date`, it is assumed that the
  vector consists of the Month and Year.
If a split vector of 1 is passed to `unknown_date`, it is assumed that the
  vector consists only of the Year.
For cases in which the year is not provided or the month is unknown, it is
  suggested that these be recoded before using this function.
}
\examples{
earliest_date(2019, 1, 0)
earliest_date(2000)
earliest_date(2019, NULL, NULL)
earliest_date(2019, NULL, 2)
earliest_date(2019, NULL, 31)

latest_date(2019)
latest_date(2019, NULL, NULL)
latest_date(2019, 0)
latest_date(2019, NULL, 2)
latest_date(2019, 2)
latest_date(2016, 2)

x <- "UN UNK 2019"
unknown_date(x)
unknown_date(x, format = "dmy")
unknown_date(x, format = "dmy", possible = "l")

unknown_date(c("01 JAN 1996", "Feb 2010", "2019"), "dmy")
unknown_date(c("01 JAN 1996", "Feb 2016"), "dmy", "latest")
}
