% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{possible_date}
\alias{possible_date}
\alias{unknown_date}
\alias{earliest_date}
\alias{latest_date}
\title{Inexact dates}
\usage{
unknown_date(
  x,
  format = "ymd",
  possible = c("earliest", "latest"),
  invalid_date_string = "^UNK?|NA$"
)

earliest_date(year, month = NULL, day = NULL, as_date = TRUE)

latest_date(year, month = NULL, day = NULL, as_date = TRUE)
}
\arguments{
\item{x}{a date "string" (see details)}

\item{format}{A date "format" (see details)}

\item{possible}{Whether to look at earliest or latest}

\item{invalid_date_string}{A string (passed as a regular expression; case insensitive)
to be removed from `x`}

\item{year}{Year}

\item{month}{Month}

\item{day}{Day}

\item{as_date}{Logical, if TRUE will return the output as a date}
}
\description{
Incomplete dates cannot be directly coerced into date values in R.
These functions help
}
\details{
`unknown_date` is a vectorized function which implements the non-vectorized
  `earliest_date` and `latest_date`.
If a split vector of 2 is passed to `unknown_date`, it is assumed that the
  vector consists of the Month and Year.
If a split vector of 1 is passed to `unknown_date`, it is assumed that the
  vector consists only of the Year.
For cases in which the year is not provided or the month is unknown, it is
  suggested that these be recoded before using this function.

The `as_date` parameter is used to allow the `unknown_date` function to work
  with a character vector and change the date class at the end
}
\examples{
earliest_date(2019, 1, 0)
earliest_date(2000)
earliest_date(2019, NULL, NULL)
earliest_date(2019, NULL, 2)
earliest_date(2019, NULL, 31)

latest_date(2019)
latest_date(2019, NULL, NULL)
latest_date(2019, 0)
latest_date(2019, NULL, 2)
latest_date(2019, 2)
latest_date(2016, 2)

x <- "UN UNK 2019"
unknown_date(x) # NA_date_

unknown_date(x, format = "dmy")
unknown_date(x, format = "dmy", possible = "l")

unknown_date(c("01 JAN 1996", "Feb 2010", "2019"), "dmy")
unknown_date(c("01 JAN 1996", "Feb 2016"), "dmy", "latest")
unknown_date("2015", possible = "e")
unknown_date("2015", possible = "l")
}
