% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{Quality check}
\usage{
qc(target, reference, ...)
}
\arguments{
\item{target}{Target}

\item{reference}{Reference}

\item{...}{Additional arguments passed to methods}
}
\description{
Perform a QC on an object
}
