% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindex.R
\name{reindex}
\alias{reindex}
\title{Re-index}
\usage{
reindex(x, index = NULL, new_index, keep_all = FALSE, ...)
}
\arguments{
\item{x}{A data.frame or named vector}

\item{index}{The name of the column to reindex by}

\item{new_index}{The new order of the index as a vector}

\item{keep_all}{Logical, whether to use only the index values in the
`new_index` or to use all available index values}

\item{...}{Additional arguments passed to methods}
}
\description{
Re-index a column based on a new index
}
\details{
When implementing the `keep_all` argument, NA values in the same class as
  the data.frame (df) are assigned.
The `index` and `new_index` must be not have any duplicates.
}
\examples{
iris1 <- head(iris, 5) \%>\% tibble::as_tibble()
iris1$index <- 1:5
reindex(iris1, "index", seq(2, 8, 2))
reindex(iris1, "index", seq(2, 8, 2), keep_all = TRUE)
}
