% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Reverse an object}
\usage{
reverse(x, ...)
}
\arguments{
\item{x}{A vector, data.frame, or matrix}

\item{...}{Additional arguments sent to methods (not used)}
}
\description{
Reverses the order of an object or a data.frame
}
\examples{
reverse(letters)
head(iris)
reverse(iris) \%>\% tail()
}
