% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-boolean.R
\name{to_boolean}
\alias{to_boolean}
\title{To boolean}
\usage{
to_boolean(x, true, false = NULL, names = FALSE)
}
\arguments{
\item{x}{A non-boolean vector}

\item{true}{A condition in which to evaluate as TRUE}

\item{false}{An optional condition to evlauate as FALSE}

\item{names}{Logical.  If TRUE, the result will print out with a name}
}
\description{
Conditionally transforms a non-boolean vector into a boolean vector
}
\examples{
x <- c("Y", "Y", "N", "Y", "N", "N/A", "Y - not sure", "YN", "YY")
to_boolean(x, "Y")
to_boolean(x, "Y", "N", names = TRUE)
}
